% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.summary.survfit.R
\name{as_flextable.summary.survfit}
\alias{as_flextable.summary.survfit}
\title{as_flextable.summary.survfit}
\usage{
\method{as_flextable}{summary.survfit}(x, which = c("n.risk", "surv"), ...)
}
\arguments{
\item{x}{returned object from function \code{survival:::summary.survfit}}

\item{which}{..}

\item{...}{..}
}
\description{
as_flextable.summary.survfit
}
\keyword{internal}
