% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable_quantile_survfit.R
\name{as_flextable_quantile_survfit}
\alias{as_flextable_quantile_survfit}
\title{Convert \link[survival]{quantile.survfit} to \link[flextable]{flextable}}
\usage{
as_flextable_quantile_survfit(x, ..., fmt = "\%.1f", units = units.survfit(x))
}
\arguments{
\item{x}{a \link[survival]{survfit} object}

\item{...}{parameters of function \link[survival]{quantile.survfit}}

\item{fmt}{see function \link[base]{sprintf}}

\item{units}{(optional) \link[base]{character} scalar, unit of time in \code{x}}
}
\description{
..
}
\note{
The confidence interval of median survival time is indeed defined by
the intersection of confidence bands of Kaplan-Meier curve and the 50\% survival horizontal line.
See \strong{Details} of function \link[survival]{print.survfit}.

Function \link[survival]{quantile.survfit} returns a \link[base]{list}.
}
\examples{
survfit(Surv(time, status) ~ 1, data = aml) |>
 as_flextable_quantile_survfit(units = 'days') 

survfit(Surv(time, status) ~ x, data = aml) |>
 as_flextable_quantile_survfit(units = 'days') 

}
\seealso{
\link[survival]{quantile.survfit},
\link[survival]{median.survfit}
}
\keyword{internal}
