% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{autolayer.summary.survfit}
\alias{autolayer.summary.survfit}
\title{Layers of Kaplan-Meier Curve of \link[survival]{summary.survfit} using \CRANpkg{ggplot2}}
\usage{
\method{autolayer}{summary.survfit}(object, ribbon = TRUE, labels = NULL, oneminus = FALSE, ...)
}
\arguments{
\item{object}{\link[survival]{summary.survfit}}

\item{ribbon}{\link[base]{logical} scalar, default \code{TRUE}}

\item{labels}{(optional) \link[base]{character} \link[base]{vector}}

\item{oneminus}{\link[base]{logical} scalar}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
