% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{autolayer.survfit}
\alias{autolayer.survfit}
\title{Layers of Kaplan-Meier Curve of \link[survival]{survfit.object} using \CRANpkg{ggplot2}}
\usage{
\method{autolayer}{survfit}(object, ...)
}
\arguments{
\item{object}{a \link[survival]{survfit.object}}

\item{...}{additional parameters of function \code{\link[=autolayer.summary.survfit]{autolayer.summary.survfit()}}}
}
\description{
..
}
\note{
It's very difficulty to grab the time \link[base]{units} of the response of a
\link[survival]{survfit} or \link[survival]{summary.survfit} object.
The only way is \code{eval(object$call$formula[[2L]], envir = eval(object$call$data))},
which is very vulnerable as \code{object$call$data} might be defined in some other internal functions.
Therefore we leave the specification of \code{xlab} (ideally the time unit) to the end user.
}
\keyword{internal}
