% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{autoplot.Surv}
\alias{autoplot.Surv}
\title{Kaplan-Meier Curve of \link[survival]{Surv} Objects using \CRANpkg{ggplot2}}
\usage{
\method{autoplot}{Surv}(object, ...)
}
\arguments{
\item{object}{\link[survival]{Surv} object}

\item{...}{..}
}
\description{
..
}
\examples{
autoplot(with(veteran, Surv(time, status))) # mask for font
}
