% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{autoplot.survfit}
\alias{autoplot.survfit}
\title{Kaplan-Meier Curve of \link[survival]{survfit.object} using \CRANpkg{ggplot2}}
\usage{
\method{autoplot}{survfit}(object, ...)
}
\arguments{
\item{object}{\link[survival]{survfit.object}}

\item{...}{additional parameters of function \code{\link[=autolayer.survfit]{autolayer.survfit()}}}
}
\description{
..
}
\seealso{
\code{survival:::plot.survfit} \code{survival:::quantile.survfit}
}
