% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitlist.R
\name{autoplot.survfitlist}
\alias{autoplot.survfitlist}
\title{autoplot.survfitlist}
\usage{
\method{autoplot}{survfitlist}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{..}
}
\description{
autoplot.survfitlist
}
\keyword{internal}
