% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_.R
\name{censor_}
\alias{censor_}
\title{Censor a \link[survival]{Surv} Object}
\usage{
censor_(x, censor)
}
\arguments{
\item{x}{\link[survival]{Surv} object}

\item{censor}{\link[base]{numeric} scalar, censoring time}
}
\description{
..
}
\details{
In practice, we sometimes want to ignore events happening after a time point.
}
\examples{
x = lung |> with(expr = Surv(time = time, event = (status == 2)))
head(x, n = 10)
x1 = censor_(x, censor = 400)
head(x1, n = 10)
}
\keyword{internal}
