% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit.R
\name{clogit}
\alias{clogit}
\title{\link[survival]{clogit} Objects}
\description{
\link[survival]{clogit} Objects
}
\examples{
# see ?survival::clogit
resp = levels(survival::logan$occupation)
n = nrow(logan)
indx = rep(1:n, length(resp))
logan2 = data.frame(logan[indx,], id = indx, tocc = factor(rep(resp, each=n)))
logan2$case = (logan2$occupation == logan2$tocc)
m = clogit(case ~ tocc + tocc:education + strata(id), data = logan2)
library(ecip); list(
 '`clogit`' = m
) |> fastmd::render2html()

}
