% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit.R
\name{clogit_S3}
\alias{clogit_S3}
\alias{endpoint.clogit}
\alias{estnm.clogit}
\alias{desc_.clogit}
\title{Additional S3 methods for \link[survival]{clogit} Objects}
\usage{
\method{endpoint}{clogit}(x)

\method{estnm}{clogit}(x)

\method{desc_}{clogit}(x)
}
\arguments{
\item{x}{\link[survival]{clogit} object}
}
\description{
Additional S3 methods for \link[survival]{clogit} Objects
}
\examples{
# see ?survival::clogit
data(logan, package = 'survival')
resp = levels(logan$occupation)
n = nrow(logan)
indx = rep(1:n, length(resp))
logan2 = data.frame(logan[indx,], id = indx, tocc = factor(rep(resp, each=n)))
logan2$case = (logan2$occupation == logan2$tocc)
(m = clogit(case ~ tocc + tocc:education + strata(id), data = logan2))
endpoint.clogit(m)

}
\keyword{internal}
