% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.zph.R
\name{cox_zph}
\alias{cox_zph}
\title{Message on Proportional Hazards Assumption Test}
\usage{
cox_zph(fit, global = TRUE, ...)
}
\arguments{
\item{fit, global, ...}{parameters of function \link[survival]{cox.zph}}
}
\description{
..
}
\details{
Function \link{cox_zph} prints ...
}
\examples{
# ?coxphw::coxphw examples
data(gastric, package = 'coxphw')
coxph(Surv(time, status) ~ radiation, data = gastric) |> cox_zph()
}
\references{
\url{http://www.sthda.com/english/wiki/cox-model-assumptions}
}
