% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme.R
\name{coxme}
\alias{coxme}
\title{\link[coxme]{coxme} Object}
\description{
\link[coxme]{coxme} Object
}
\examples{
library(coxme)
m = coxme(Surv(time, status) ~ ph.ecog + age + (1|inst), data = lung) 
library(ecip); list(
 '`coxme`' = m
) |> fastmd::render2html()
}
