% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{coxph}
\alias{coxph}
\title{\link[survival]{coxph} Objects}
\description{
\link[survival]{coxph} Objects
}
\examples{
# ?survival::cox.zph
m = ovarian |>
 within.data.frame(expr = {
  ecog.ps = factor(ecog.ps)
 }) |>
 coxph(formula = Surv(futime, fustat) ~ age + ecog.ps)

library(ecip); list(
 '`coxph`' = m
) |> fastmd::render2html()

}
