% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.coxph.R
\name{diagnose.coxph}
\alias{diagnose.coxph}
\title{diagnose.coxph}
\usage{
diagnose.coxph(
  object,
  type = c("GrambschTherneau", "CoxSnell", "deviance", "schoenfeld", "dfbetas",
    "martingale", "dev_by_x"),
  ...
)
}
\arguments{
\item{object}{..}

\item{type}{..}

\item{...}{..}
}
\description{
diagnose.coxph
}
\examples{
(c1 = coxph(Surv(time, status == 2) ~ age + sex + ph.ecog, data = lung))
diagnose.coxph(c1, type = 'GrambschTherneau')
diagnose.coxph(c1, type = 'CoxSnell')
#diagnose(c1, type = 'martingale') # debug this!!! could due to missing data
# the waiting time should enter the proportional hazards model nonlinearly
diagnose.coxph(c1, type = 'deviance') # mask for font
#diagnose(c1, type = 'dev_by_x') # debug this!!! could due to missing data
diagnose.coxph(c1, type = 'schoenfeld') # mask for font
# diagnose(c1, type = 'dfbetas') # ggscatter deprecated
}
