% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.survfit.R
\name{diagnose.survfit}
\alias{diagnose.survfit}
\title{diagnose.survfit}
\usage{
diagnose.survfit(object, type = "survival", ...)
}
\arguments{
\item{object}{..}

\item{type}{..}

\item{...}{..}
}
\description{
..
}
\examples{
f1 = survfit(Surv(time, status == 2) ~ ph.ecog, data = lung)
#grid::grid.draw(diagnose(f1, type = 'survival'))
diagnose.survfit(f1, type = 'hazard')
}
\references{
\url{http://www.ats.ucla.edu/stat/examples/asa/test_proportionality.htm}
}
