% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{.pval.survdiff}
\alias{.pval.survdiff}
\title{Get \eqn{p}-values from \link[survival]{survdiff} Object}
\usage{
.pval.survdiff(x)
}
\arguments{
\item{x}{\link[survival]{survdiff} object}
}
\description{
Get \eqn{p}-values from \link[survival]{survdiff} Object
}
\examples{
m = survdiff(Surv(futime, fustat) ~ rx,data=ovarian)
.pval.survdiff(m)

}
