% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitcox.R
\name{endpoint.survfitcox}
\alias{endpoint.survfitcox}
\title{Endpoint of \code{'survfitcox'} Object}
\usage{
\method{endpoint}{survfitcox}(x)
}
\arguments{
\item{x}{an object of class \code{'survfitcox'}, returned from function \code{survival:::survfit.coxph()}}
}
\description{
Endpoint of \code{'survfitcox'} Object
}
\examples{
coxph(Surv(time, status) ~ celltype, data = veteran) |> 
  survfit() |>
  endpoint.survfitcox()
}
\keyword{internal}
