% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{getTime.Surv}
\alias{getTime.Surv}
\title{Get Time from \link[survival]{Surv} Object}
\usage{
getTime.Surv(x, type = c("event", "censor", "any"), ...)
}
\arguments{
\item{x}{\link[survival]{Surv} objects}

\item{type}{\link[base]{character} scalar}

\item{...}{potential parameters, currently not in use}
}
\description{
..
}
\note{
Do not clash with S3 \link[stats]{time}!
}
