% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKM.R
\name{ggKM}
\alias{ggKM}
\alias{ggKM.formula}
\alias{ggKM.coxph}
\title{Kaplan Meier Curve via \CRANpkg{ggplot2}}
\usage{
ggKM(object, data, ...)

\method{ggKM}{formula}(formula, data, ..., envir = parent.frame())

\method{ggKM}{coxph}(object, ...)
}
\arguments{
\item{object}{..}

\item{data}{\link[base]{data.frame}}

\item{...}{additional parameters of function \code{\link[=autolayer.survfit]{autolayer.survfit()}}}

\item{formula}{\link[stats]{formula}}

\item{envir}{..}
}
\value{
Function \code{\link[=ggKM]{ggKM()}} returns a \link[ggplot2]{ggplot} object.
}
\description{
..
}
\details{
Function \code{\link[=ggKM]{ggKM()}} is more elaborated than \code{\link[=autoplot.survfit]{autoplot.survfit()}}.
}
\references{
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}

For function \code{\link[=ggKM]{ggKM()}} \url{http://www.stat.unipg.it/luca/R/CumIncidence.R}

Package \pkg{ggfortify} \url{https://cran.r-project.org/web/packages/ggfortify/}
}
