% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKM_rpart.R
\name{ggKM.rpart}
\alias{ggKM.rpart}
\title{Kaplan-Meier Curves based on \link[rpart]{rpart}}
\usage{
\method{ggKM}{rpart}(object, ...)
}
\arguments{
\item{object}{\link[rpart]{rpart}}

\item{...}{..}
}
\description{
..
}
\examples{
library(rpart)
rp = rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE)
library(rmd.tzh); list(
  'ggKM.rpart' = ggKM(rp)
) |> render_(file = 'ggKM.rpart')

}
\keyword{internal}
