% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplot.R
\name{ggsurvplot}
\alias{ggsurvplot}
\title{\link[survminer]{ggsurvplot}}
\description{
\link[survminer]{ggsurvplot}
}
\examples{
p = survfit(Surv(time, status) ~ sex, data = lung) |>
 survminer::ggsurvplot(data = lung)
list(
 '`ggsurvplot`' = p
) |> fastmd::render2html(file = 'ggsurvplot')

}
