% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplot_rpart.R
\name{ggsurvplot_rpart}
\alias{ggsurvplot_rpart}
\title{Kaplan-Meier Curves of \link[rpart]{rpart} via \CRANpkg{survminer}}
\usage{
ggsurvplot_rpart(fit, ...)
}
\arguments{
\item{fit}{\link[rpart]{rpart}}

\item{...}{..}
}
\description{
Kaplan-Meier curves based on recursive partitioning and regression trees \link[rpart]{rpart}.
}
\examples{
rp = rpart::rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE)
list(
  'ggsurvplot_rpart' = rp |> ggsurvplot_rpart(pval.coord = c(750, .95))
) |> rmd.tzh::render_(file = 'ggsurvplot_rpart')

}
\keyword{internal}
