% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{hms}
\alias{hms}
\alias{hms.character}
\alias{hms.POSIXlt}
\alias{hms.POSIXct}
\title{Hour-Minute-Second}
\usage{
hms(x, ...)

\method{hms}{character}(x, format = "\%H:\%M:\%OS", ...)

\method{hms}{POSIXlt}(x, ...)

\method{hms}{POSIXct}(x, ...)
}
\arguments{
\item{x}{a \link[base]{POSIXlt} or \link[base]{POSIXct} object}

\item{...}{..}

\item{format}{see function \link[base]{as.POSIXlt.character}}
}
\value{
The function \code{\link[=hms]{hms()}} returns a \link[base]{difftime} object.
}
\description{
The function \link[readxl]{read_excel} might read a date-less time stamp
(e.g., \verb{3:00:00 PM} as \verb{1899-12-31 15:00:00 UTC}).
}
\examples{
f = tempfile(fileext = '.xlsx')
data.frame(
 id = c('A', 'B'),
 tm = c('8:30:00 AM', '3:00:00 PM')
) |>
 writexl::write_xlsx(path = f)

x = f |>
 readxl::read_excel()
x$tm |> class() # 'character'
x$tm |> hms() |> suppressWarnings()

}
\seealso{
\link[base]{ISOdatetime};
\code{lubridate:::year.default};
\code{lubridate:::tz.POSIXt}
}
\keyword{internal}
