% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{is.Surv.endpoint}
\alias{is.Surv.endpoint}
\title{\code{S3} Generic Function \code{\link[=is.Surv.endpoint]{is.Surv.endpoint()}}}
\usage{
is.Surv.endpoint(object)
}
\arguments{
\item{object}{an R object}
}
\description{
An \code{S3} generic function for downstream packages,
to determine if an R object is a model with a \link[survival]{Surv} endpoint
}
\keyword{internal}
