% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{md_.coxph}
\alias{md_.coxph}
\title{R Markdown Lines for \link[survival]{coxph}}
\usage{
\method{md_}{coxph}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[survival]{coxph}
}
\examples{
?survival::cox.zph
m = ovarian |>
 within.data.frame(expr = {
  ecog.ps = factor(ecog.ps)
 }) |>
 coxph(formula = Surv(futime, fustat) ~ age + ecog.ps)

library(ecip); 
list('`coxph`' = m) |> rmd.tzh::render_(file = 'coxph')

}
\keyword{internal}
