% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplot_rpart.R
\name{md_.ggsurvplot}
\alias{md_.ggsurvplot}
\title{md_.ggsurvplot}
\usage{
\method{md_}{ggsurvplot}(x, xnm, ...)
}
\arguments{
\item{x}{..}

\item{xnm}{..}

\item{...}{..}
}
\description{
md_.ggsurvplot
}
\keyword{internal}
