% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{md_.survdiff}
\alias{md_.survdiff}
\title{Fast Markdown Lines for \link[survival]{survdiff} Object}
\usage{
\method{md_}{survdiff}(x, xnm, ...)
}
\arguments{
\item{x}{a \link[survival]{survdiff} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
list(
 logrank = survdiff(Surv(time, status) ~ x, data = aml)
) |> fastmd::render_(file = 'logrank')
}
\keyword{internal}
