% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{md_.survfit}
\alias{md_.survfit}
\title{md_.survfit}
\usage{
\method{md_}{survfit}(x, xnm, ...)
}
\arguments{
\item{x}{\link[survival]{survfit.object}}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
list(
 'no strata' = survfit(Surv(time, status) ~ 1, data = aml),
 'one strata' = survfit(Surv(time, status) ~ x, data = aml)
) |> rmd.tzh::render_(file = 'survfit')
}
\keyword{internal}
