% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitlist.R
\name{md_.survfitlist}
\alias{md_.survfitlist}
\title{\code{\link[=md_.survfitlist]{md_.survfitlist()}}}
\usage{
\method{md_}{survfitlist}(x, xnm, ...)
}
\arguments{
\item{x}{a \code{'survfitlist'} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
