% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{model_desc.survdiff}
\alias{model_desc.survdiff}
\title{model_desc.survdiff}
\usage{
model_desc.survdiff(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{..}
}
\value{
..
}
\description{
model_desc.survdiff
}
