% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitlist.R
\name{msurvfit}
\alias{msurvfit}
\title{Kaplan Meier Curves of Multiple Endpoints}
\usage{
msurvfit(formula, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}.
E.g. \code{c(y1, y2) ~ x1 + x2} for
an array of Kaplan Meier curves of \code{y1 ~ x1 + x2} and \code{y2 ~ x1 + x2}}

\item{...}{additional parameters of function \link[survival]{survfit.formula}}
}
\value{
Function \code{\link[=msurvfit]{msurvfit()}} returns a \code{'survfitlist'}.
}
\description{
Array of Kaplan-Meier curves of multiple endpoints and same predictors.
}
\keyword{internal}
