% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{nobs.survdiff}
\alias{nobs.survdiff}
\title{Sample Size of \link[survival]{survdiff} Object}
\usage{
\method{nobs}{survdiff}(object, ...)
}
\arguments{
\item{object}{\link[survival]{survdiff} object}

\item{...}{additional parameters, currently not in use}
}
\description{
Sample Size of \link[survival]{survdiff} Object
}
\examples{
m = survdiff(Surv(futime, fustat) ~ rx,data=ovarian)
stopifnot(nobs.survdiff(m) == nrow(ovarian))
}
