% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{nobsText_survfit}
\alias{nobsText_survfit}
\alias{nobsText.survfit}
\alias{nobsText.summary.survfit}
\title{Sample Size of \link[survival]{survfit.object}}
\usage{
\method{nobsText}{survfit}(x)

\method{nobsText}{summary.survfit}(x)
}
\arguments{
\item{x}{a \link[survival]{survfit.object} or \link[survival]{summary.survfit} object}
}
\description{
Sample Size of \link[survival]{survfit.object}
}
\examples{
library(ecip)
survfit(Surv(time, status) ~ x, data = aml) |> 
 nobsText()
survfit(Surv(time, status) ~ x, data = aml) |> 
 summary() |>
 nobsText()
}
\keyword{internal}
