% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneminus.survfit.R
\name{oneminus}
\alias{oneminus}
\alias{oneminus.survfit}
\alias{oneminus.summary.survfit}
\title{One-Minus Kaplan-Meier Estimates}
\usage{
oneminus.survfit(x)

oneminus.summary.survfit(x)
}
\arguments{
\item{x}{..}
}
\description{
One-Minus Kaplan-Meier Estimates
}
\keyword{internal}
