% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_survfit_test.R
\name{pairwise_survfit_test}
\alias{pairwise_survfit_test}
\title{pairwise_survfit_test}
\usage{
pairwise_survfit_test(
  x,
  times = stop("must provide time point"),
  units = attr(x, which = "units", exact = TRUE),
  ...
)
}
\arguments{
\item{x}{\link[survival]{survfit.object}}

\item{times}{\link[base]{numeric} scalar or \link[base]{vector}, see parameter \code{times} of
\code{survival:::summary.survfit}}

\item{units}{\link[base]{character} scalar, the time unit}

\item{...}{..}
}
\description{
..
}
\details{
..
}
\examples{
library(survival)
x = survfit(Surv(stop-start, event = (status!=0)) ~ treatment, 
  data = bladder1)
pairwise_survfit_test(x, times = c(10, 20))

}
\seealso{
\link[stats]{pairwise.table}
}
