% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSIXct2difftime.R
\name{POSIXct2difftime}
\alias{POSIXct2difftime}
\title{Convert No-Date \link[base]{POSIXct} to \link[base]{difftime}}
\usage{
POSIXct2difftime(
  x,
  else_return = stop("Expecting all '1899-12-31 hr:min:sec UTC'")
)
}
\arguments{
\item{x}{\link[base]{POSIXct}}

\item{else_return}{exception handling}
}
\description{
..
}
\details{
The function \link[readxl]{read_excel} reads \code{'hour:min:sec'} as \code{'1899-12-31 hr:min:sec UTC'}
}
\note{
The function \link[base]{as.difftime} is \strong{not} an S3 generic function.
}
\seealso{
\code{lubridate:::year.default};
\code{lubridate:::tz.POSIXt}
}
\keyword{internal}
