% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme.R
\name{S3_coxme}
\alias{S3_coxme}
\alias{.pval.summary.coxme}
\alias{nobsText.coxme}
\alias{desc_.coxme}
\alias{estnm.coxme}
\title{S3 methods for \link[coxme]{coxme}}
\usage{
\method{.pval}{summary.coxme}(x)

\method{nobsText}{coxme}(x)

\method{desc_}{coxme}(x)

\method{estnm}{coxme}(x)
}
\arguments{
\item{x}{\link[coxme]{coxme} object}
}
\description{
S3 methods for \link[coxme]{coxme}
}
\examples{
library(coxme)
m = coxme(Surv(time, status) ~ ph.ecog + age + (1|inst), data = lung) 
m |> nobsText.coxme()
}
\keyword{internal}
