% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.zph.R
\name{S3_coxzph}
\alias{S3_coxzph}
\alias{.pval.cox.zph}
\alias{desc_.cox.zph}
\title{S3 methods of \link[survival]{cox.zph}}
\usage{
.pval.cox.zph(x)

desc_.cox.zph(x)
}
\arguments{
\item{x}{a \link[survival]{cox.zph} object}
}
\description{
S3 methods of \link[survival]{cox.zph}
}
\examples{
?survival::cox.zph
m = coxph(Surv(futime, fustat) ~ age + ecog.ps, data = ovarian) |> cox.zph() 
.pval.cox.zph(m)
}
