% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{S3_survdiff}
\alias{S3_survdiff}
\alias{nobsText.survdiff}
\alias{.pval.survdiff}
\title{S3 Methods \link[survival]{survdiff} Object}
\usage{
\method{nobsText}{survdiff}(x)

\method{.pval}{survdiff}(x)
}
\arguments{
\item{x}{\link[survival]{survdiff} object}
}
\description{
S3 Methods \link[survival]{survdiff} Object
}
\note{
Function \code{?broom:::nobs.survdiff} (2024-09-26) is \strong{WRONG}.
}
\examples{
m = survdiff(Surv(futime, fustat) ~ rx, data = ovarian) 
m |> nobsText.survdiff()
m |> .pval.survdiff()
 
}
\keyword{internal}
