% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{S3_survfit}
\alias{S3_survfit}
\alias{conf_level.survfit}
\alias{nobsText.survfit}
\title{S3 Methods for \link[survival]{survfit} Objects}
\usage{
conf_level.survfit(x)

nobsText.survfit(x)
}
\arguments{
\item{x}{\link[survival]{survfit} object}
}
\description{
S3 Methods for \link[survival]{survfit} Objects
}
\examples{
m = survfit(Surv(time, status) ~ x, data = aml)
m |> conf_level.survfit()
m |> nobsText.survfit()

}
