% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitcox.R
\name{S3_survfitcox}
\alias{S3_survfitcox}
\alias{endpoint.survfitcox}
\title{Additional S3 methods for \code{survfitcox} Objects}
\usage{
endpoint.survfitcox(x)
}
\arguments{
\item{x}{\code{survfitcox} object}
}
\description{
Additional S3 methods for \code{survfitcox} Objects
}
