% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{S3_survreg}
\alias{S3_survreg}
\alias{.pval.summary.survreg}
\title{S3 methods for \link[survival]{survreg}}
\usage{
.pval.summary.survreg(x)
}
\arguments{
\item{x}{a \link[survival]{survreg} or \link[survival]{summary.survreg}}
}
\description{
S3 methods for \link[survival]{survreg}
}
\examples{
library(survival)
# ?survival::survreg
m = survreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, dist = 'weibull', scale = 1) 

}
