% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_doc.R
\docType{data}
\name{slight_mod}
\alias{slight_mod}
\alias{lung2}
\alias{rotterdam2}
\title{Slight Modification of Datasets in \CRANpkg{survival} Package}
\format{
..

An object of class \code{data.frame} with 2982 rows and 17 columns.
}
\usage{
lung2

rotterdam2
}
\description{
..
}
\keyword{datasets}
