% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_ggKM.R
\name{subset_ggKM}
\alias{subset_ggKM}
\title{subset_ggKM}
\usage{
subset_ggKM(formula, data, ids, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{data}{\link[base]{data.frame}}

\item{ids}{\link[base]{list} of \link[base]{integer} or \link[base]{logical} \link[base]{vector},
indices for subset of \code{data}}

\item{...}{..}
}
\description{
Produce a batch of Kaplan-Meier curves for \emph{subset analysis}.
}
\examples{
dim(colon1 <- subset(colon, etype == 2)) # death
colon1a = within(colon1, expr = {
 time = as.difftime(time, units = 'days')
 OS = Surv(time, status)
})
ids = with(colon1a, list(All = TRUE, Female = (sex == 0), Male = (sex == 1)))
subset_ggKM(OS ~ rx, ids = ids, data = colon1a)
}
