% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitlist.R
\name{subset_survfit}
\alias{subset_survfit}
\title{subset_survfit}
\usage{
subset_survfit(formula, subset, data, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{subset}{\link[base]{language} to indicate
the subset(s) of \code{data}}

\item{data}{\link[base]{data.frame}}

\item{...}{..}
}
\value{
Function \code{\link[=subset_survfit]{subset_survfit()}} returns a \code{'survfitlist'}.
}
\description{
Multiple \link[survival]{survfit} in different subset.
}
\keyword{internal}
