% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{survdiff}
\alias{survdiff}
\title{\link[survival]{survdiff} Object}
\description{
\link[survival]{survdiff} Object
}
\note{
The \code{S3} method \code{broom:::nobs.survdiff()} (2026-01-26) is \strong{WRONG}.
}
\examples{
list(
 logrank = survdiff(Surv(time, status) ~ x, data = aml),
 '$\\\\rho=1$' = survdiff(Surv(time, status) ~ x, data = aml, rho = 1),
 '$0<\\\\rho<1$' = survdiff(Surv(time, status) ~ x, data = aml, rho = .5)
) |> fastmd::render2html()
}
