% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff.R
\name{survdiff_}
\alias{survdiff_}
\title{S3 Generic Function \code{\link[=survdiff_]{survdiff_()}}}
\usage{
survdiff_(object, ...)
}
\arguments{
\item{object}{an R object}

\item{...}{additional parameters}
}
\description{
An \code{S3} generic function for downstream packages
}
\keyword{internal}
