% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survdiff_rpart.R
\name{survdiff_rpart}
\alias{survdiff_rpart}
\title{Test Survival Curve Differences based on \link[rpart]{rpart}}
\usage{
survdiff_rpart(object, ...)
}
\arguments{
\item{object}{\link[rpart]{rpart}}

\item{...}{..}
}
\description{
..
}
