% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.rpart.R
\name{survfit.rpart}
\alias{survfit.rpart}
\title{Survival Curves based on \link[rpart]{rpart}}
\usage{
\method{survfit}{rpart}(formula, ...)
}
\arguments{
\item{formula}{\link[rpart]{rpart}}

\item{...}{..}
}
\description{
..
}
\examples{
library(rpart)
rp = rpart(Surv(time, status) ~ age, data = veteran, maxdepth = 2L, model = TRUE)
rp |> survfit.rpart()

}
