% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{survival_bib}
\alias{survival_bib}
\alias{.kaplan_meier58}
\alias{.cox72}
\alias{.ripatti04}
\alias{.therneau03}
\title{\link[utils]{bibentry} for Survival Analysis}
\usage{
.kaplan_meier58()

.cox72()

.ripatti04()

.therneau03()
}
\description{
\link[utils]{bibentry} for Survival Analysis
}
\keyword{internal}
