% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry.R
\name{survival_bibentry}
\alias{survival_bibentry}
\alias{KaplanMeier58}
\alias{Cox72}
\alias{Ripatti04}
\alias{Therneau03}
\title{\link[utils]{bibentry} for Survival Analysis}
\usage{
KaplanMeier58()

Cox72()

Ripatti04()

Therneau03()
}
\description{
\link[utils]{bibentry} for Survival Analysis
}
