% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survOver.R
\name{survOver}
\alias{survOver}
\title{\link[survival]{Surv}ival Over A Time}
\usage{
survOver(x, tm, censored = NA)
}
\arguments{
\item{x}{a \link[survival]{Surv} object}

\item{tm}{\link[base]{numeric} scalar}

\item{censored}{\link[base]{logical} scalar,
returned value for censored subjects}
}
\value{
Function \link{survOver} returns a \link[base]{logical} \link[base]{vector}.
}
\description{
\link[survival]{Surv}ival Over A Time
}
\examples{
(x = Surv(1:4, event = c(0,1,0,1)))
survOver(x, tm = 3)
survOver(x, tm = 3, censor = FALSE)
}
\keyword{internal}
