% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{survreg}
\alias{survreg}
\title{\link[survival]{survreg} Object}
\description{
\link[survival]{survreg} Object
}
\examples{
# ?survival::survreg
m = survreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, dist = 'weibull', scale = 1) 

}
