% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime.R
\name{timeUnits}
\alias{timeUnits}
\title{Additional Time Units \code{'months'} and \code{'years'}}
\usage{
timeUnits()
}
\value{
Function \link{timeUnits} returns a named constant \link[base]{character} \link[base]{vector}.
}
\description{
To support additional time units \code{'months'} and \code{'years'} for \link[base]{difftime} object.
}
\details{
Every 4 years has \verb{1461(=365*4+1)} days, or \verb{48(=4*12)} months.
Therefore, every month has \verb{30.44(=1461/48)} days, or \verb{4.35(=1461/48/7)} weeks.

Every year has 12 months.
}
\note{
Function \link[base]{units<-.difftime}
only supports
\code{'secs'}, \code{'mins'}, \code{'hours'}, \code{'days'}, \code{'weeks'}.
}
\keyword{internal}
