% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{units.Surv}
\alias{units.Surv}
\title{Time Unit of a \link[survival]{Surv} object}
\usage{
\method{units}{Surv}(x)
}
\arguments{
\item{x}{\link[survival]{Surv} object}
}
\description{
..
}
\seealso{
\link[base]{units}, \code{Hmisc:::units.Surv}
}
\keyword{internal}
