% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{units_survfit}
\alias{units_survfit}
\alias{units.survfit}
\alias{units.summary.survfit}
\title{Time Unit of a \link[survival]{survfit.object} or \link[survival]{summary.survfit} Object}
\usage{
\method{units}{survfit}(x)

\method{units}{summary.survfit}(x)
}
\arguments{
\item{x}{a \link[survival]{survfit.object} or \link[survival]{summary.survfit} object}
}
\description{
..
}
\seealso{
\link[base]{units}
}
\keyword{internal}
