% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.coxph.R
\name{vars_in_data}
\alias{vars_in_data}
\alias{vnumeric}
\title{Names of type in data (to be deprecated)}
\usage{
vnumeric(x)
}
\arguments{
\item{x}{\link[base]{data.frame}}
}
\description{
..
}
\details{
\link{vnumeric} ..
}
