% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{as_flextable.summary.survfit}
\alias{as_flextable.summary.survfit}
\title{Convert \link[survival]{summary.survfit} Object to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{summary.survfit}(x, which = c("surv", "n.risk"), ...)
}
\arguments{
\item{x}{a \link[survival]{summary.survfit} object}

\item{which}{..}

\item{...}{..}
}
\description{
Convert \link[survival]{summary.survfit} Object to \link[flextable]{flextable}
}
\keyword{internal}
