% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{autolayer.survfit}
\alias{autolayer.survfit}
\title{Layers of Kaplan-Meier Curve of \link[survival]{survfit.object} using \CRANpkg{ggplot2}}
\usage{
\method{autolayer}{survfit}(object, times, ribbon = TRUE, labels = NULL, ...)
}
\arguments{
\item{object}{\link[survival]{survfit.object}}

\item{times}{(optional) \link[base]{numeric} scalar or \link[base]{vector},
time points where survival rates as well as the confidence intervals are plotted}

\item{ribbon}{\link[base]{logical} scalar, default \code{TRUE}}

\item{labels}{(optional) \link[base]{character} \link[base]{vector}}

\item{...}{..}
}
\description{
..
}
\note{
It's very difficulty to grab the time \link[base]{units} of the response of a
\link[survival]{survfit} object.
The only way is \code{eval(object$call$formula[[2L]], envir = eval(object$call$data))},
which is very vulnerable as \code{object$call$data} might be defined in some other internal functions.
Therefore we leave the specification of \code{xlab} (ideally the time unit) to the end user.
}
