% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfitcox.R
\name{autoplot.survfitcox}
\alias{autoplot.survfitcox}
\title{Kaplan-Meier Curve of \code{survfitcox} Objects using \CRANpkg{ggplot2}}
\usage{
\method{autoplot}{survfitcox}(object, ...)
}
\arguments{
\item{object}{\code{'survfitcox'} object, returned from \code{survival:::survfit.coxph}}

\item{...}{..}
}
\description{
..
}
\examples{
m1 = coxph(Surv(time, status) ~ celltype, data = veteran)
m1 |> survfit() |> autoplot()
m1 |> survfit(newdata = veteran[1:90, ]) |> autoplot()
m1 |> survfit(newdata = veteran[1:90, ]) |> autoplot(times = c(250))
}
\keyword{internal}
