% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.survfit.R
\name{fortify.survfit}
\alias{fortify.survfit}
\title{Convert \link[survival]{survfit.object} to \link[base]{data.frame}}
\usage{
\method{fortify}{survfit}(model, ..., units = "", oneminus = FALSE)
}
\arguments{
\item{model}{\link[survival]{survfit.object}.
\code{'survfitcox'} object returned from
\code{survival:::survfit.coxph} is also supported, as of 2022-09-19}

\item{...}{additional parameters of \link[base]{as.data.frame.list}}

\item{units}{\link[base]{character}}

\item{oneminus}{\link[base]{logical} scalar}
}
\value{
Function \code{\link[=fortify.survfit]{fortify.survfit()}} returns a \code{survival:::summary.survfit} object.
}
\description{
..
}
\details{
A column named \code{'strata'}
of all-equal values \code{'all_subjects'}
for \link[survival]{survfit.object} without a strata.
}
\keyword{internal}
