% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggKM.R
\name{ggKM}
\alias{ggKM}
\alias{ggKM.formula}
\alias{ggKM.coxph}
\title{Kaplan Meier Curve via \CRANpkg{ggplot2}}
\usage{
ggKM(object, data, ...)

\method{ggKM}{formula}(formula, data, ..., envir = parent.frame())

\method{ggKM}{coxph}(object, ...)
}
\arguments{
\item{object}{..}

\item{data}{\link[base]{data.frame}}

\item{...}{additional parameters of function \link{autolayer.survfit}}

\item{formula}{\link[stats]{formula}}

\item{envir}{..}
}
\value{
Function \code{\link[=ggKM]{ggKM()}} returns a \link[ggplot2]{ggplot} object.
}
\description{
..
}
\details{
Function \code{\link[=ggKM]{ggKM()}} is more elaborated than \link{autoplot.survfit}.
}
\examples{
# ?veteran
# https://www.medrxiv.org/content/10.1101/2022.05.22.22275430v1
veteran = within(veteran, expr = {
  time = as.difftime(time, units = 'days')
  os = Surv(time = time, event = status)
})
ggKM(os ~ 1, data = veteran)
ggKM(os ~ celltype, data = veteran)
ggKM(os ~ celltype, data = veteran, times = c(100))
ggKM(os ~ celltype, data = veteran, labels = letters[1:4])
ggKM(os ~ celltype + I(age > 60), data = veteran)
}
\references{
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}

For function \code{\link[=ggKM]{ggKM()}} \url{http://www.stat.unipg.it/luca/R/CumIncidence.R}

Package \pkg{ggfortify} \url{https://cran.r-project.org/web/packages/ggfortify/}
}
