% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_ggKM.R
\name{m_ggKM}
\alias{m_ggKM}
\title{Kaplan Meier Curves of Multiple Endpoints}
\usage{
m_ggKM(formula, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}.
E.g. \code{cbind(y1, y2) ~ x1 + x2 + x3} for
an array of Kaplan Meier curves of \code{y1 ~ x1 + x2 + x3} and \code{y2 ~ x1 + x2 + x3}}

\item{...}{additional parameters of \code{\link[=ggKM.formula]{ggKM.formula()}}}
}
\description{
Array of Kaplan-Meier curves of multiple endpoints and same predictors.
}
\examples{
# y-axis not the same
p = ggKM(rfs ~ chemo, data = rotterdam2)
p = ggKM(os ~ chemo, data = rotterdam2)
p = m_ggKM(cbind(rfs, os) ~ chemo, data = rotterdam2)
}
