% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{md_.summary.survfit}
\alias{md_.summary.survfit}
\title{Fast Markdown Lines for \link[survival]{summary.survfit} Object}
\usage{
\method{md_}{summary.survfit}(x, xnm, ...)
}
\arguments{
\item{x}{a \link[survival]{summary.survfit} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\keyword{internal}
