% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{md_.survfit}
\alias{md_.survfit}
\title{Fast Markdown Lines for \link[survival]{survfit.object}}
\usage{
\method{md_}{survfit}(x, xnm, ...)
}
\arguments{
\item{x}{\link[survival]{survfit.object}}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
s0 = survfit(Surv(time, status) ~ 1, data = aml)
s1 = survfit(Surv(time, status) ~ x, data = aml)
list(
 'no strata' = list(s0, summary(s0, times = c(10, 25))),
 'one strata' = list(s1, summary(s1, times = c(10, 25)))
) |> fastmd::render2html()
}
\keyword{internal}
