% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R, R/asDifftime.R, R/survfit.R
\name{more_units<-.Surv}
\alias{more_units<-.Surv}
\alias{more_units_set}
\alias{more_units<-}
\alias{more_units<-.difftime}
\alias{more_units<-.survfit}
\alias{more_units<-.summary.survfit}
\title{Use Month and Year as \link[base]{units}}
\usage{
\method{more_units}{Surv}(x) <- value

more_units(x) <- value

\method{more_units}{difftime}(x) <- value

\method{more_units}{survfit}(x) <- value

\method{more_units}{summary.survfit}(x) <- value
}
\arguments{
\item{x}{see \strong{Usage}}

\item{value}{\link[base]{character} scalar,
choice of unit, e.g.,
\itemize{
\item{\code{'secs'}, \code{'mins'}, \code{'hours'}, \code{'days'}, \code{'weeks'},
which are also supported in the function \link[base]{units<-.difftime}}
\item{\code{'months'} and \code{'years'}, which are only supported by the function \link{more_units<-}}
}}
}
\value{
The \code{S3} method \link{more_units<-.Surv} returns a \link[survival]{Surv} object.

The \code{S3} method \link{more_units<-.difftime} returns a \link[base]{difftime} object.

The \code{S3} method \link{more_units<-.survfit} returns a \link[survival]{survfit.object}.

The \code{S3} method \link{more_units<-.summary.survfit} returns a \link[survival]{summary.survfit} object.
}
\description{
Set \link[base]{units} of R objects,
with additional support of \code{'months'} and \code{'years'}.
}
\examples{
(x = Sys.Date() - as.Date('2021-01-01'))
tryCatch(units(x) <- 'months', error = identity)
more_units(x) = 'months'; x
more_units(x) = 'years'; x

}
\keyword{internal}
