% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{S3_coxph}
\alias{S3_coxph}
\alias{nobsText.coxph}
\alias{desc_.coxph}
\alias{estnm.coxph}
\alias{.pval.summary.coxph}
\title{S3 methods of \link[survival]{coxph}}
\usage{
nobsText.coxph(x)

desc_.coxph(x)

estnm.coxph(x)

.pval.summary.coxph(x)
}
\arguments{
\item{x}{an \link[survival]{coxph} object}
}
\description{
S3 methods of \link[survival]{coxph}
}
\examples{
?survival::cox.zph
m = ovarian |>
 within.data.frame(expr = {
  ecog.ps = factor(ecog.ps)
 }) |>
 coxph(formula = Surv(futime, fustat) ~ age + ecog.ps)
survival:::nobs.coxph(m) # number of events!!
# MuMIn:::nobs.coxph(m) # also number of events
nobsText.coxph(m)
}
