% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{S3_coxph}
\alias{S3_coxph}
\alias{nobsText.coxph}
\alias{desc_.coxph}
\alias{estnm.coxph}
\alias{.pval.summary.coxph}
\title{S3 methods of \link[survival]{coxph}}
\usage{
\method{nobsText}{coxph}(x)

\method{desc_}{coxph}(x)

\method{estnm}{coxph}(x)

\method{.pval}{summary.coxph}(x)
}
\arguments{
\item{x}{an \link[survival]{coxph} object}
}
\description{
S3 methods of \link[survival]{coxph}
}
\keyword{internal}
