% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{S3_survfit}
\alias{S3_survfit}
\alias{nobsText.survfit}
\title{S3 Methods for \link[survival]{survfit} Objects}
\usage{
nobsText.survfit(x)
}
\arguments{
\item{x}{\link[survival]{survfit} or \link[survival]{summary.survfit} object}
}
\description{
S3 Methods for \link[survival]{survfit} Objects
}
\examples{
survfit(Surv(time, status) ~ x, data = aml) |> 
 nobsText.survfit()
}
