% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_ggKM.R
\name{subset_ggKM}
\alias{subset_ggKM}
\title{subset_ggKM}
\usage{
subset_ggKM(formula, subset, data, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}}

\item{subset}{\link[base]{language} to indicate
the subset(s) of \code{data}}

\item{data}{\link[base]{data.frame}}

\item{...}{..}
}
\value{
Function \code{\link[=subset_ggKM]{subset_ggKM()}} returns a \link[patchwork]{patchwork}.
}
\description{
Produce a \link[patchwork]{patchwork} of Kaplan-Meier curves for \emph{subset analysis}.
}
\examples{
p = subset_ggKM(os ~ ph.ecog, subset = list(
  All = TRUE, Female = (sex == 'Female'), Male = (sex == 'Male')
), data = lung2)

id = quote(list(
  All = TRUE, Female = (sex == 'Female'), Male = (sex == 'Male')
))
p = subset_ggKM(os ~ ph.ecog, subset = id, data = lung2)
}
