% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{survival_bib}
\alias{survival_bib}
\alias{.kaplan_meier58}
\alias{.cox72}
\alias{.ripatti04}
\alias{.therneau03}
\title{\link[utils]{bibentry} for Survival Analysis}
\usage{
.kaplan_meier58(key = "KaplanMeier58", ...)

.cox72(key = "Cox72", ...)

.ripatti04(key = "Ripatti04", ...)

.therneau03(key = "Therneau03", ...)
}
\arguments{
\item{key, ...}{parameters of function \link[utils]{bibentry}}
}
\description{
\link[utils]{bibentry} for Survival Analysis
}
\keyword{internal}
